Refactor the social login button to reduce duplication
