+++ packages/api-core/src/me-social-logins/types.ts
@@ -21,7 +21,7 @@ export interface ConnectSocialUserArgs {
 }
 
 export interface PostLoginRequestBodyObj {
-	service: 'github' | 'google' | 'apple' | 'paypal';
+	service: 'google' | 'apple' | 'github' | 'paypal';
 	redirect_uri?: string;
 	state?: string;
 	auth_code: string;
